requirements.txt
psycopg2-binary==2.7.5
SQLAlchemy==1.2.12
scrapy==1.5.1
scrapy-selenium==0.0.5


